requirements.txt
anyio==4.9.0
attrs==23.2.0
cachetools==5.5.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.4.1
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.6
joblib==1.4.2
numpy==2.1.1
oauthlib==3.2.2
opencv-python==4.10.0.84
outcome==1.3.0.post0
pillow==10.4.0
proto-plus==1.26.1
protobuf==6.30.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pyparsing==3.2.1
PySocks==1.7.1
python-telegram-bot==22.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.17.2
setuptools==69.5.1
sniffio==1.3.0
sortedcontainers==2.4.0
threadpoolctl==3.6.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
wsproto==1.2.0
